Move data files to src/assets/data
