+++ src/assets/data/courses.js
