+++ src/assets/data/education.js
